pyproject.toml
[project]
name = "plan-trib"
version = "0.1.0"
description = "Sistema de planejamento tributário comparativo (Simples Nacional × Lucro Presumido × Lucro Real)"
authors = [{ name = "GUINER CONTABILIDADE", email = "guiner@guiner.com" }]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "argon2-cffi>=23.1.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.17",
    "google-genai>=0.3.0",
    "httpx>=0.27.0",
    "structlog>=24.4.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "respx>=0.21.0",
    "syrupy>=4.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
